requirements.txt
Flask
Flask-Cors
psycopg2-binary
pandas
gunicorn
